work/requirements.txt
apache-airflow>=2.5.0
google-cloud-aiplatform>=1.34.0
requests>=2.28.0
pendulum>=2.1.2

